Moved enums into a seperate file.
