+++ srv/game.js
@@ -1,18 +1,7 @@
 const idGen = require('./utils/idgen');
 const shuf = require('./utils/shuffle');
 
-const states = {
-	LOBBY: 'Lobby', //Not started the game yet.
-	PLAYING: 'Playing', //In-game.
-	POST: 'Post-game' //After the game, so we can show people a map with live locations on it.
-}
-
-const roles = {
-	FUGITIVE: 'fugitive',
-	HUNTER: 'hunter',
-	SPECTATOR: 'spectator',
-	POSTGAME: 'postgame'
-}
+const { states, roles } = require('./utils/enums');
 
 class Game {
