+++ srv/utils/enums.js
@@ -1,0 +1,17 @@
+const states = {
+	LOBBY: 'Lobby', //Not started the game yet.
+	PLAYING: 'Playing', //In-game.
+	POST: 'Post-game' //After the game, so we can show people a map with live locations on it.
+}
+
+const roles = {
+	FUGITIVE: 'fugitive',
+	HUNTER: 'hunter',
+	SPECTATOR: 'spectator',
+	POSTGAME: 'postgame'
+}
+
+module.exports = {
+    states: states,
+    roles: roles
+};
